+++ api/src/main/java/kr/co/pawpaw/api/config/SecurityConfig.java
@@ -60,7 +60,7 @@ public class SecurityConfig {
                 .antMatchers(HttpMethod.GET, "/api/auth/sign-up/check/duplicate/email").permitAll()
                 .antMatchers(HttpMethod.POST, "/api/auth/change/password/mail").permitAll()
                 .antMatchers(HttpMethod.POST, "/api/auth/change/password").permitAll()
-                .antMatchers(HttpMethod.GET, "/api/user/email").permitAll()
+                .antMatchers(HttpMethod.GET, "/api/auth/email").permitAll()
                 .antMatchers("/swagger-ui.html").permitAll()
                 .antMatchers("/oauth2/**").permitAll()
                 .anyRequest().authenticated())
